Refactor ManagerMessager into IPC
