+++ tests/common/mqtt/test_wrapper.py
@@ -8,7 +8,7 @@ import pytest
 from pydantic import BaseModel
 
 from astoria.common.config import MQTTBrokerInfo
-from astoria.common.messages.base import ManagerMessage
+from astoria.common.ipc import ManagerMessage
 from astoria.common.mqtt.topic import Topic
 from astoria.common.mqtt.wrapper import MQTTWrapper
